requirements.txt
python-docx
pdf2image
Pillow
pytesseract
fpdf2
odfpy
striprtf
beautifulsoup4
markdown
